fix bug repeat right answer if you press end game while you already have right answer

+++ utilities.py
@@ -1,3 +1,4 @@
+import json
 import random
 
 import telegram
@@ -32,18 +33,26 @@ def new_question(chat_id, r_db, quiz_questions):
     questions = list(quiz_questions.keys())
     random_question_number = random.choice(questions)
     random_question = quiz_questions[random_question_number]['question']
-    r_db.set(chat_id, random_question_number)
+    print(quiz_questions[random_question_number]['answer'])
+    user_data = json.dumps({
+        'current_question_id': random_question_number,
+        'is_answer_correct': False
+    })
+    r_db.set(chat_id, user_data)
     return random_question
 
 
 def end_quiz(chat_id, r_db, quiz_questions):
-    current_question_id = r_db.get(chat_id)
-    if current_question_id:
-        current_question = quiz_questions[current_question_id]['question']
-        current_question_true_answer = quiz_questions[current_question_id]['answer']
+    current_question_id = json.loads(r_db.get(chat_id))
+    if current_question_id and current_question_id['is_answer_correct'] is False:
+        current_question = quiz_questions[current_question_id['current_question_id']]['question']
+        current_question_true_answer = quiz_questions[current_question_id['current_question_id']]['answer']
         text = f'Текущий вопрос: {current_question}\n\nОтвет: {current_question_true_answer}'
         r_db.delete(chat_id)
         return {'status': True, 'data': text}
+    elif current_question_id['current_question_id'] and current_question_id['is_answer_correct']:
+        text = TEXTS['next_question']
+        return {'status': False, 'data': text}
     else:
         text = TEXTS['quiz_not_started']
         return {'status': False, 'data': text}
@@ -60,7 +69,7 @@ def get_user_score(chat_id, r_db):
 
 
 def get_user_last_question(chat_id, r_db, quiz_questions):
-    current_question_id = r_db.get(chat_id)
+    current_question_id = json.loads(r_db.get(chat_id))
     if current_question_id:
         question_id_number = current_question_id.lower().replace("вопрос ", "")
         current_question = quiz_questions[current_question_id]['question']
@@ -71,13 +80,18 @@ def get_user_last_question(chat_id, r_db, quiz_questions):
 
 
 def waiting_for_question_answer(chat_id, user_message, r_db, quiz_questions):
-    current_question_id = r_db.get(chat_id)
-    if current_question_id:
-        true_answer = quiz_questions[current_question_id]['answer'].lower()
+    current_question_id = json.loads(r_db.get(chat_id))
+    if current_question_id['current_question_id']:
+        true_answer = quiz_questions[current_question_id['current_question_id']]['answer'].lower()
         user_text = user_message.lower()
         case_is_true = _is_answer_correct(user_text, true_answer)
         if case_is_true:
             answer_to_user = f"{TEXTS['true_answer']} {TEXTS['next_question']}"
+            user_data = json.dumps({
+                'current_question_id': current_question_id['current_question_id'],
+                'is_answer_correct': True
+            })
+            r_db.set(chat_id, user_data)
         else:
             answer_to_user = f"{TEXTS['false_answer']} {TEXTS['try_again']}"
         return {'status': True, 'true_answer': case_is_true, 'data': answer_to_user, 'answer': true_answer}
